# app/api-node/swagger.yaml
openapi: 3.0.0
info:
    description: 'BattleBots API'
    license:
        name: MIT
    title: BattleBots
    version: 0.0.1
paths:
...
    '/api/games/start/{id}':
        put:
            operationId: GameStart
            responses:
                '200':
                    description: 'game started'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '404':
                    description: 'game not found'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
    '/api/games/stop/{id}':
        put:
            operationId: GameStop
            responses:
                '200':
                    description: 'game stopped'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '404':
                    description: 'game not found'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
    /api/games/create:
        post:
            operationId: GameCreate
            responses:
                '201':
                    description: 'game created'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            requestBody:
                content:
                    appplication/json;charset=UTF-8:
                        schema:
                            $ref: '#/components/schemas/IGameResource'
                required: true
    '/api/games/end/{id}':
        put:
            operationId: GameEnd
            responses:
                '200':
                    description: 'game ended'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '404':
                    description: 'game not found'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
    /api/games/:
        get:
            operationId: GameList
            responses:
                '200':
                    description: 'game list'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
        put:
            operationId: GameUpdate
            responses:
                '200':
                    description: 'game updated'
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        'application/json; charset=UTF-8':
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            requestBody:
                content:
                    'application/json; charset=UTF-8':
                        schema:
                            $ref: '#/components/schemas/IGameResource'
                required: true
    '/api/games/{id}':
        get:
            operationId: GameDetail
            responses:
                '200':
                    description: 'game detail'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '404':
                    description: 'game not found'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
        delete:
            operationId: GameDelete
            responses:
                '200':
                    description: 'game deleted'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '204':
                    description: 'No content'
                '400':
                    description: ""
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
                '404':
                    description: 'game not found'
                    content:
                        application/json;charset=UTF-8:
                            schema:
                                $ref: '#/components/schemas/HttpResponseModelIGameResource'
            description: ""
            security:
                -
                    Bearer:
                        - ROLE_USER
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    schema:
                        type: number
                        format: double
...
servers:
    -
        url: /
components:
...
    securitySchemes:
        Bearer:
            type: http
            scheme: bearer
    schemas:
        IArenaResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                arena_name:
                    type: string
                    description: ""
                available:
                    type: number
                    format: double
                    description: ""
                bots:
                    type: array
                    items:
                        $ref: '#/components/schemas/IBotsResource'
                    description: ""
            type: object
        IBotsResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                botIp:
                    type: string
                    description: ""
                running:
                    type: number
                    format: double
                    description: ""
                taken:
                    type: number
                    format: double
                    description: ""
                name:
                    type: string
                    description: ""
                speed:
                    type: number
                    format: double
                    description: ""
                damage:
                    type: number
                    format: double
                    description: ""
                fireRate:
                    type: number
                    format: double
                    description: ""
                armor:
                    type: number
                    format: double
                    description: ""
                streams:
                    type: array
                    items:
                        $ref: '#/components/schemas/IStreamResource'
                    description: ""
                games:
                    type: array
                    items:
                        $ref: '#/components/schemas/IGameResource'
                    description: ""
                gameProfile:
                    $ref: '#/components/schemas/IGameProfileResource'
            type: object
            required:
                - id
                - botIp
                - running
                - taken
                - name
                - speed
                - damage
                - fireRate
                - armor
        IStreamResource:
            description: ""
            properties:
                game:
                    $ref: '#/components/schemas/IGameResource'
                robot:
                    $ref: '#/components/schemas/IBotsResource'
                kinesisUrl:
                    type: string
                    description: ""
                s3Url:
                    type: string
                    description: ""
                private:
                    type: number
                    format: double
                    description: ""
                running:
                    type: number
                    format: double
                    description: ""
                duration:
                    type: number
                    format: double
                    description: ""
                encodage:
                    type: string
                    description: ""
                id:
                    type: number
                    format: double
                    description: ""
            type: object
            required:
                - kinesisUrl
                - s3Url
                - private
                - running
                - duration
                - encodage
        IGameResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                name:
                    type: string
                    description: ""
                arena:
                    $ref: '#/components/schemas/IArenaResource'
                bots:
                    type: array
                    items:
                        $ref: '#/components/schemas/IBotsResource'
                    description: ""
                streams:
                    type: array
                    items:
                        $ref: '#/components/schemas/IStreamResource'
                    description: ""
                status:
                    enum:
                        - CREATED
                        - STARTED
                        - STOPPED
                        - ENDED
                    type: string
                    description: ""
                createdAt:
                    type: string
                    description: ""
                updatedAt:
                    type: string
                    description: ""
                token:
                    type: string
                    description: ""
            type: object
            required:
                - name
                - status
        IGameProfileResource:
            description: ""
            properties:
                total_points:
                    type: number
                    format: double
                    description: ""
                id:
                    type: number
                    format: double
                    description: ""
                name:
                    type: string
                    description: ""
                user:
                    $ref: '#/components/schemas/IUserResource'
            type: object
        IUserResource:
            description: ""
            properties:
                id:
                    type: number
                    format: double
                    description: ""
                firstname:
                    type: string
                    description: ""
                lastname:
                    type: string
                    description: ""
                pseudo:
                    type: string
                    description: ""
                email:
                    type: string
                    description: ""
                roles:
                    type: string
                    description: ""
                address:
                    type: string
                    description: ""
                createdAt:
                    type: string
                    description: ""
                updatedAt:
                    type: string
                    description: ""
                password:
                    type: string
                    description: ""
                gameProfile:
                    type: array
                    items:
                        $ref: '#/components/schemas/IGameProfileResource'
                    description: ""
            type: object
            required:
                - pseudo
                - email
                - roles
...
        HttpResponseModelIGameResource:
            description: ""
            properties:
                message:
                    type: string
                    description: ""
                httpCode:
                    type: number
                    format: double
                    description: ""
                data:
                    $ref: '#/components/schemas/IGameResource'
            type: object
...